refactor: extract shared RecipeCard component for home and recipes pages
Card styles and getDaysOld helper moved from Recipes/Index into a new
RecipeCard client component (src/components/RecipeCard/). Home page
now uses it so cards are visually identical on both pages.
