+++ src/views/Recipes/recipes.scss
@@ -249,101 +249,6 @@
     }
   }
 
-  .recipe-cards {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-    gap: 14px;
-  }
-
-  .recipe-card {
-    display: flex;
-    gap: 12px;
-    padding: clamp(.8rem, 1.5vw, 1rem);
-    border-radius: 10px;
-    border: 1px solid var(--vscode-border);
-    background: var(--vscode-panel);
-    transition: border-color 140ms ease, background-color 140ms ease, box-shadow 140ms ease;
-
-    &:hover {
-      border-color: var(--vscode-accent);
-      background: var(--accent-hover);
-      box-shadow: 0 4px 12px var(--shadow-container);
-    }
-
-    &.is-selected {
-      border-color: var(--vscode-accent);
-      background: var(--accent-hover-strong);
-      box-shadow: 0 0 0 1px var(--vscode-accent);
-    }
-  }
-
-  .card-checkbox {
-    display: flex;
-    align-items: flex-start;
-    padding-top: 2px;
-  }
-
-  .card-content {
-    flex: 1;
-    min-width: 0;
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-    text-decoration: none;
-    color: inherit;
-  }
-
-  .card-header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: 8px;
-    flex-wrap: nowrap;
-  }
-
-  .card-title {
-    margin: 0;
-    font-size: clamp(.75rem, 2.5vw, 1rem);
-    font-weight: 600;
-    color: var(--vscode-text);
-    flex: 1;
-    min-width: 0;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-
-  .card-badges {
-    display: flex;
-    align-items: center;
-    gap: 6px;
-    flex-shrink: 0;
-    flex-wrap: nowrap;
-  }
-
-  .card-description {
-    margin: 0;
-    font-size: clamp(.6rem, 2vw, .875rem);
-    color: var(--vscode-muted);
-    line-height: 1.5;
-    display: -webkit-box;
-    -webkit-line-clamp: 2;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-  }
-
-  .card-footer {
-    display: flex;
-    align-items: center;
-    gap: 14px;
-    margin-top: auto;
-  }
-
-  .card-meta {
-    font-size: 12px;
-    color: var(--vscode-muted);
-  }
-
   @media (max-width: 720px) {
     .recipes-header {
       flex-direction: column;
